+++ js/packages/client-websocket/src/index.ts
@@ -73,19 +73,19 @@ function startSyncInWorker(
     switch (msg._tag) {
       case "db_change":
         // todo: test this ignoreNotif logic so it doesn't break (e.g., by new microtask enqueueing) in the future.
-        ignoreNotif = true;
         // tell rx about the changes from a different connection
-        args.rx.__internalNotifyListeners(msg.collectedChanges);
-        ignoreNotif = false;
+        args.rx.__internalNotifyListeners(msg.collectedChanges, "sync");
         break;
     }
   };
 
   disposables.push(
-    args.rx.onAny(() => {
-      // ignore changes coming from the sync layer itself?
-      if (ignoreNotif) {
-        console.log("ignoring changes from sync layer itself");
+    args.rx.onAny((updates, src) => {
+      // ignore changes coming from the sync layer itself.
+      // This is because RX calls into sync layer telling it there was a change.
+      // But if change came from sync layer  itself, we don't need to tell sync layer this.
+      if (src !== "thisTab") {
+        console.log("ignoring changes from sync layer / other tabs");
         return;
       }