+++ js/packages/direct-connect-nodejs/src/config/DefaultConfig.ts
@@ -1,0 +1,12 @@
+import { Config } from "../Types.js";
+
+const DefaultConfig: Config = {
+  serviceName: "vlcn-default",
+  dbsDir: "./dbs",
+  cacheTtlInSeconds: 60 * 5,
+  notifyLatencyInMs: 10,
+  serviceDbPath: "./dbs/service.db",
+  msgContentType: "application/json",
+};
+
+export default DefaultConfig;