+++ js/packages/sandbox/vite.config.js
@@ -5,10 +5,10 @@ import { defineConfig } from "vite";
 export default defineConfig({
   base: "./",
   build: {
-    target: "es2020",
+    target: "esnext",
     rollupOptions: {
       input: {
-        page1: resolve(__dirname, "page1.html"),
+        page1: resolve(__dirname, "direct-connect-browser.html"),
         page2: resolve(__dirname, "page2.html"),
       },
     },
@@ -16,7 +16,7 @@ export default defineConfig({
 
   optimizeDeps: {
     esbuildOptions: {
-      target: "es2020",
+      target: "esnext",
     },
   },
 });