+++ js/packages/sandbox/CHANGELOG.md
@@ -1,5 +1,40 @@
 # sandbox
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @vlcn.io/direct-connect-browser@0.0.5
+
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @vlcn.io/direct-connect-browser@0.0.4
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @vlcn.io/direct-connect-browser@0.0.3
+
+## 0.1.1
+
+### Patch Changes
+
+- fts5, sqlite 3.42.1, direct-connect packages
+- Updated dependencies [6dbfdcb]
+- Updated dependencies
+  - @vlcn.io/crsqlite-wasm@0.10.1
+  - @vlcn.io/direct-connect-browser@0.0.2
+  - @vlcn.io/direct-connect-common@0.1.1
+  - @vlcn.io/direct-connect-nodejs@0.1.1
+  - @vlcn.io/react@0.10.1
+  - @vlcn.io/rx-tbl@0.9.1
+
 ## 0.1.0
 
 ### Minor Changes